Create io_test_py2.py
This is a Windows example of how to create a Pandas DataFrame in Python 2 and then store it using REDIS so that you can read it into a Python 3 code. The reason for this is a company X who has not created Py3 drivers for their hardware. In order to use the hardware, you are stuck with Py2, but in reality, for GUIs and UX, you would like to be using Py3... REDIS is running in Ubuntu under WSL2 in Windows. 
Use these links:
https://github.com/redis/redis-om-python
https://redis.io/docs/getting-started/installation/install-redis-on-windows/